kernelsnap: add Silex WiFi module support
The driver source code for this module is external to the kernel tree
location:
https://github.com/boundarydevices/qcacld-2.0

However, since all the modules *need* to be part of the kernelsnap due
to current Snappy implementation, we need to integrate the driver
inside the tree.

So we now retrieve the latest source code as archive to speed up the
process, avoiding to retrieve the whole git tree.

Signed-off-by: Laci Tele <laci@boundarydevices.com>
Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>