snappy: add prod profile when generating image
To be used as follows:
$ PROD=1 make snappy

This will use the nitrogen gadget and kernel from the store instead of
the local ones. That allows that image to use 'snap refresh' on those
packages.

Also remove the image size parameter as Ubuntu now expands the partition
to the storage size during the first bootup.

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>